tests: Move snapshot_restore_common to common folder
Extract the inline mod snapshot_restore_common from
integration.rs into common/snapshot_restore_common.rs.
This keeps it accessible as a shared module since other
modules reference its pub(crate) functions.
